8. Minor Updates (Performance Improvements and Layered *.svg Exports)
Improvements to the search function which takes clicks and associates them with an isometric cell. Instead of searching the entire grid, the Y Coordinate of the click is extracted and processed to only search the relevant row. 

Layered SVG's are also introduced, which put the Auxetics, and the Outline of the cut on separate layers, such that they are cut in that respective order. This should reduce the chance of the cut moving during the cutting process, and allows users to remove the outline, with compatible laser cutters.